deal better with unexpected indents
before, the recovery algorithm would try to repair an unexpected indent
like this:

print 1
    a = 1

--->

print 1
try:
    a = 1

because that is the shortest sequence of inserts that allows it to
continue. But then, the "except" is missing, causing a spurious second
error.

Instead, add a recovery rule to the grammar, that makes it possible to
start a suite in any place with a single insert.